+++ app.js
@@ -3,7 +3,7 @@ import 'regenerator-runtime/runtime';
 import express from 'express';
 import dotenv from 'dotenv';
 import logger from 'morgan';
-import { roleRoutes, userRoutes, documentRoutes } from './routes';
+import { roleRoutes, userRoutes, documentRoutes } from './src/routes';
 
 dotenv.config();
