+++ package.json
@@ -12,8 +12,8 @@
     "db:migrate:testdb": "NODE_ENV=test node_modules/.bin/sequelize db:migrate",
     "undo:migrations": "node_modules/.bin/sequelize db:migrate:undo:all",
     "undo:migrations:testdb": "NODE_ENV=test npm run undo:migrations",
-    "prebuild": "yarn clean && mkdir build && cp -r config build",
-    "build": "babel ./ --ignore 'build,node_modules,test,coverage,seeders,tmp' -s -d build",
+    "prebuild": "yarn clean && mkdir build",
+    "build": "babel src --out-dir build --copy-files",
     "clean": "rm -rf build"
   },
   "repository": {
